Refactor: create MultipleUserSelector component
